+++ lib/fl_extended.dart
@@ -8,7 +8,7 @@ export 'src/builder/listenable_builder.dart';
 /// components
 export 'src/components/animation.dart';
 export 'src/components/unified_button.dart';
-export 'src/components/icon_box.dart';
+export 'src/components/icon_label.dart';
 export 'src/components/fl_inherited.dart';
 export 'src/components/progress_indicator.dart';